Use `XCTExpectFailure` instead of skipping tests in scheme
My concern with using the scheme (or a Test Plan) to skip tests is that
the setting is tucked away behind multiple UI layers and mouse clicks. I
worry that we'll forget about it.

By still running these two tests and marking their failures as expected,
we make the issue more visible.